Add option to only recognize Swiss German Input (#3)
Added a toggle with the options "Schweizerdeutsch" and "Automatisch". It controls whether the speech service should only recognize Swiss German or automatically recognize the four default languages German, English, French, Spanish.
Usecase: Chat with Bot
