+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+## Unreleased
+
+### Added
+- Option to recognize only Swiss German input. (#3)
+
+## v1.0.0 - 2022-01-26
+### Initial Release
+Try out the new Swiss Natural Language Processing Voice "Leni".
+
+Follow the instructions [here](https://github.com/isolutionsag/cognitive-services-speech-demo/tree/v1.0.0) to get started.
