+++ src/App.tsx
@@ -27,10 +27,10 @@ import BingSearchConfig from "./models/BingSearchConfig";
 import UseCase, { UseCaseModels } from "./util/UseCase";
 import GravityItemsArea from "./components/common/GravityItemsArea";
 import UseCaseTemplate from "./components/usecases/UseCaseTemplate";
-import FourLangToSwissTranslation from "./components/usecases/FourLangToSwissTranslation";
-import ChatWithBot from "./components/usecases/ChatWithBot";
-import RealtimeTranscription from "./components/usecases/RealtimeTranscription";
-import NewsReader from "./components/usecases/NewsReader";
+import FourLangToSwissTranslation from "./components/usecases/four_lang_to_swiss_translation/FourLangToSwissTranslation";
+import ChatWithBot from "./components/usecases/chat_with_bot/ChatWithBot";
+import RealtimeTranscription from "./components/usecases/realtime_transcription/RealtimeTranscription";
+import NewsReader from "./components/usecases/news_reader/NewsReader";
 import { ArrowBack, VpnKey } from "@mui/icons-material";
 import { ThemeProvider, createTheme } from "@mui/material/styles";
 import AppBar from "@mui/material/AppBar";
